+++ src/cpp11.cpp
@@ -13,17 +13,17 @@ extern "C" SEXP _readODScpp_ods_get_sheet_names(SEXP file, SEXP include_external
   END_CPP11
 }
 // read_ods_.cpp
-cpp11::strings read_ods_(const std::string file, int start_row, int stop_row, int start_col, int stop_col, const int sheet);
-extern "C" SEXP _readODScpp_read_ods_(SEXP file, SEXP start_row, SEXP stop_row, SEXP start_col, SEXP stop_col, SEXP sheet) {
+cpp11::strings read_ods_(const std::string file, int start_row, int stop_row, int start_col, int stop_col, const int sheet, const bool formula_as_formula);
+extern "C" SEXP _readODScpp_read_ods_(SEXP file, SEXP start_row, SEXP stop_row, SEXP start_col, SEXP stop_col, SEXP sheet, SEXP formula_as_formula) {
   BEGIN_CPP11
-    return cpp11::as_sexp(read_ods_(cpp11::as_cpp<cpp11::decay_t<const std::string>>(file), cpp11::as_cpp<cpp11::decay_t<int>>(start_row), cpp11::as_cpp<cpp11::decay_t<int>>(stop_row), cpp11::as_cpp<cpp11::decay_t<int>>(start_col), cpp11::as_cpp<cpp11::decay_t<int>>(stop_col), cpp11::as_cpp<cpp11::decay_t<const int>>(sheet)));
+    return cpp11::as_sexp(read_ods_(cpp11::as_cpp<cpp11::decay_t<const std::string>>(file), cpp11::as_cpp<cpp11::decay_t<int>>(start_row), cpp11::as_cpp<cpp11::decay_t<int>>(stop_row), cpp11::as_cpp<cpp11::decay_t<int>>(start_col), cpp11::as_cpp<cpp11::decay_t<int>>(stop_col), cpp11::as_cpp<cpp11::decay_t<const int>>(sheet), cpp11::as_cpp<cpp11::decay_t<const bool>>(formula_as_formula)));
   END_CPP11
 }
 
 extern "C" {
 static const R_CallMethodDef CallEntries[] = {
     {"_readODScpp_ods_get_sheet_names", (DL_FUNC) &_readODScpp_ods_get_sheet_names, 2},
-    {"_readODScpp_read_ods_",           (DL_FUNC) &_readODScpp_read_ods_,           6},
+    {"_readODScpp_read_ods_",           (DL_FUNC) &_readODScpp_read_ods_,           7},
     {NULL, NULL, 0}
 };
 }